pyproject.toml
[project]
name = "chatdata-mcp-server"
version = "0.1.0"
description = "MCP服务器，提供多种工具功能"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx>=0.24.0",
    "black>=23.3.0",
    "autopep8>=2.0.0",
    "psycopg2-binary>=2.9.5",
    "tabulate>=0.9.0",
    "bs4>=0.0.2",
    "numpy>=2.2.3",
    "sympy>=1.13.3",
    "pyyaml>=6.0.2",
    "xmltodict>=0.14.2",
]

[project.scripts]
server = "server.server:main"
